requirements.txt
discord.py
fastapi[all]
python-dotenv
asyncpg
aiofiles
